Fixed incomplete PDF page when rerendering is required
I (Alex) mixed up the order of a few lines that are supposed to detect
when rerendering a page is necessary due to the page not being fully
rendered. So now, even if multiple rerenders are necessary to get a
fully rendered page, it should still display the full page.